requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pymongo==4.6.1
python-dotenv==1.0.1
requests==2.31.0
wikipedia==1.4.0
groq==0.34.1
jinja2==3.1.3
pydantic==2.7.0
